pyproject.toml
[tool.poetry]
name = "meetpi"
version = "0.1.0"
description = "Google Calendar meeting status indicator for Raspberry Pi + Blinkt"
authors = ["Dave Hunt <dave.hunt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
blinkt = { version = "^0.1.2", markers = "sys_platform == 'linux'" }
google-api-python-client = "^1.12.8"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.2"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
flake8 = "^3.8.4"
pytest-cov = "^2.11.1"
flake8-docstrings = "^1.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
